Move EndBlocker to dedicated file
